Use more points per edge in the B and LoS grids
Use 3, 5, 7, 9 points per edge in the B and LoS grids, in order
to have more resolution when plotting the grids. Also plot the
magnetic unit vector in each point of the grid.

Do not plot the central point (center of the UCD star), as there
are 0 division errors and mathematical singularities that affect
the computations.

Use odd numbers when defining number of points per edge of the
grid, in order to have always a point in the center of the star,
that is removed afterwards.