Compute specific intensities of rays and plane Y'Z'
Compute specific intensities of each of the rays parallel to the
LoS direction. Compute the total specific intensity in the plane
Y'Z' perpendicular to the LoS (x')

To do so, simulated (not real) coefficients of emission and
absorption have been used for the calculations. Also, for the moment
only B_x' (in the direction of the LoS), has been defined as influencing
the radio emission. Finally, also Ne (number density of non-thermal
electrons), has been simulated with a random number (we took Ne=1).
All of these aspects are TODOs, though to be improved in future versions.

TODO: compute the total density of flux, thanks to the already
calculated total specific intensity