Allow values for cube side till 1000
Allow values of discretization for cube side till 1000

The grid for the computations can go now to 1000, but values >51
begins to take already quite a lot of time for the computation.

Values suggested shall be odd (to have one point at the middle of
the grid), and go from 13 (really low values) to 101.