Verify that entered values for 'n' is odd
Verify that entered values for n (3D), n (2D) and n (1D) are
odd, as the software has been designed taking in mind that one
of the voxels was falling in the middle of the (sub)stellar object.

This voxel is then removed and not taken into account in the
computations, as it falls inside the star.