Use much faster approach to group voxels in rays
Use a much faster approach to regroup voxels in rays.
For it, do not remove the central voxel 0,0,0.

Group the voxels thanks to a series of for loops, taking into
account how they are organized at the beginning, when they are
added to the self.voxels array. This way, it is not
necessary to check all the voxels list as many times as the
number of voxels in the list, in order to sort them into rays.
Now the list is browsed a single time to place each of the voxels
in the corresponding ray.